Improvements

+++ src/diffusers/schedulers/scheduling_ddpm_flax.py
@@ -265,7 +265,7 @@ class FlaxDDPMScheduler(FlaxSchedulerMixin, ConfigMixin):
         # 6. Add noise
         def random_variance():
             split_key = jax.random.split(key, num=1)
-            noise = jax.random.normal(split_key, shape=model_output.shape)
+            noise = jax.random.normal(split_key, shape=model_output.shape, dtype=self.config.dtype)
             return (self._get_variance(state, t, predicted_variance=predicted_variance) ** 0.5) * noise
 
         variance = jnp.where(t > 0, random_variance(), jnp.zeros(model_output.shape, dtype=self.config.dtype))
